nothing

+++ src/main/java/LIC/UC04v1/repositories/UserRepository.java
@@ -2,9 +2,11 @@ package LIC.UC04v1.repositories;
 
 import LIC.UC04v1.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository("userRepository")
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
     User findByEmail(String email);
 }
+
